+++ e_face_x4/app/connectors/__init__.py
@@ -1,5 +1,6 @@
 from .buspro import BusproConnector
 from .etherm import EThermConnector
 from .media import EkonexMediaConnector
+from .local_media import LocalMediaConnector
 
-__all__ = ["BusproConnector", "EThermConnector", "EkonexMediaConnector"]
+__all__ = ["BusproConnector", "EThermConnector", "EkonexMediaConnector", "LocalMediaConnector"]